Refatorando CadastroMilitar e CadastroServico para utilizar novos componentes FormMilitar e FormServico.
